+++ gloc_ui/lib/pages/LoadingPage.dart
